[ANDROSDK-1957] Move service to network and update AttributeCall and Attribute
