+++ core/src/main/java/org/hisp/dhis/android/core/attribute/internal/AttributeCall.kt
@@ -35,13 +35,13 @@ import org.koin.core.annotation.Singleton
 
 @Singleton
 internal class AttributeCall(
-    private val service: AttributeService,
+    private val networkHandler: AttributeNetworkHandler,
     private val handler: AttributeHandler,
     private val apiDownloader: APIDownloader,
 ) : UidsCallCoroutines<Attribute> {
     override suspend fun download(uids: Set<String>): List<Attribute> {
         return apiDownloader.downloadPartitioned(uids, MAX_UID_LIST_SIZE, handler) { partitionUids ->
-            service.getAttributes(AttributeFields.allFields, AttributeFields.uid.`in`(partitionUids), false)
+            networkHandler.getAttributes(AttributeFields.allFields, AttributeFields.uid.`in`(partitionUids), false)
         }
     }
